+++ src/core/abc/storage.abc.ts
@@ -1,4 +1,6 @@
-export abstract class MediaStorage {
+import { SessionConfig } from '../../structures/sessions.dto';
+
+abstract class MediaStorage {
   abstract save(
     messageId: string,
     mimetype: string,
@@ -6,18 +8,33 @@ export abstract class MediaStorage {
   ): Promise<string>;
 }
 
-export abstract class LocalSessionStorage {
-  public sessionsFolder: string;
+abstract class SessionConfigRepository {
+  abstract save(sessionName: string, config: SessionConfig): Promise<void>;
 
+  abstract get(sessionName: string): Promise<SessionConfig>;
+}
+
+abstract class SessionStorage {
   protected constructor(public engine: string) {
     this.engine = engine;
   }
 
-  abstract init();
+  abstract init(): Promise<void>;
+
+  abstract clean(sessionName: string): Promise<void>;
+
+  abstract getAll(): Promise<string[]>;
+
+  //
+  // Repositories
+  //
+  abstract get configRepository(): SessionConfigRepository;
+}
+
+abstract class LocalSessionStorage extends SessionStorage {
+  public sessionsFolder: string;
 
   abstract getFolderPath(sessionName: string): string;
-
-  abstract clean(sessionName: string);
-
-  abstract getAll(): string[];
 }
+
+export { LocalSessionStorage, MediaStorage, SessionConfigRepository };