+++ src/core/storage.core.ts
@@ -1,8 +1,14 @@
-import * as fs from 'fs';
+import * as crypto from 'crypto';
+import * as fs from 'fs/promises';
 import * as os from 'os';
 import * as path from 'path';
 
-import { LocalSessionStorage, MediaStorage } from './abc/storage.abc';
+import { SessionConfig } from '../structures/sessions.dto';
+import {
+  LocalSessionStorage,
+  MediaStorage,
+  SessionConfigRepository,
+} from './abc/storage.abc';
 import { DOCS_URL } from './exceptions';
 
 export class MediaStorageCore implements MediaStorage {
@@ -17,35 +23,74 @@ export class MediaStorageCore implements MediaStorage {
   }
 }
 
+class LocalSessionConfigRepository extends SessionConfigRepository {
+  FILENAME = '.waha.session.config.json';
+
+  constructor(private storage: LocalSessionStorage) {
+    super();
+  }
+
+  async get(sessionName: string): Promise<SessionConfig> {
+    const filepath = this.getFilePath(sessionName);
+    // Check file exists
+    try {
+      await fs.access(filepath, fs.constants.F_OK);
+    } catch (error) {
+      return undefined;
+    }
+
+    const content = await fs.readFile(filepath, 'utf-8');
+    return JSON.parse(content);
+  }
+
+  async save(sessionName: string, config: SessionConfig) {
+    const filepath = this.getFilePath(sessionName);
+    const content = JSON.stringify(config);
+    await fs.writeFile(filepath, content);
+  }
+
+  private getFilePath(sessionName): string {
+    const folder = this.storage.getFolderPath(sessionName);
+    return path.join(folder, this.FILENAME);
+  }
+}
+
 export class SessionStorageCore extends LocalSessionStorage {
   public sessionsFolder: string;
 
   constructor(engine: string) {
     super(engine);
-    this.sessionsFolder = path.join(os.tmpdir(), 'waha-');
+    this.sessionsFolder = path.join(os.tmpdir(), 'waha-sessions');
   }
 
   get engineFolder() {
     return path.join(this.sessionsFolder, this.engine);
   }
 
-  init() {
-    fs.mkdirSync(this.engineFolder, { recursive: true });
+  async init() {
+    await fs.mkdir(this.engineFolder, { recursive: true });
   }
 
   getFolderPath(name: string): string {
-    return path.join(this.engineFolder, name);
+    const suffix = crypto.createHash('md5').update(name).digest('hex');
+    return path.join(this.engineFolder, `${name}-${suffix}`);
   }
 
-  clean(sessionName: string) {
+  async clean(sessionName: string) {
     const folder = this.getFolderPath(sessionName);
-    fs.rmSync(folder, { recursive: true });
+    await fs.rm(folder, { recursive: true });
   }
 
-  getAll(): string[] {
-    return fs
-      .readdirSync(this.engineFolder, { withFileTypes: true })
+  async getAll(): Promise<string[]> {
+    const content = await fs.readdir(this.engineFolder, {
+      withFileTypes: true,
+    });
+    return content
       .filter((dirent) => dirent.isDirectory())
       .map((dirent) => dirent.name);
+  }
+
+  get configRepository(): LocalSessionConfigRepository {
+    return new LocalSessionConfigRepository(this);
   }
 }